requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3

pytest>=7.4
pytest-django>=4.7

python-dotenv>=1.0
